Move core logic into separate functions
There should be no changes to command-line usage,
but now the core functionality can be used
independently of the command line (i.e. imported by other modules).
