+++ electron_inject/__init__.py
@@ -8,7 +8,11 @@ import websocket
 import json
 import socket
 import subprocess
-import time
+import os
+import sys
+import logging
+
+logger = logging.getLogger(__name__)
 
 SCRIPT_HOTKEYS_F12_DEVTOOLS_F5_REFRESH = """document.addEventListener("keydown", function (e) {
     if (e.which === 123) {
@@ -118,6 +122,56 @@ class ElectronRemoteDebugger(object):
                 break
             time.sleep(1)
         return cls("localhost", port=port)
+
+
+def launch_url(url):
+    #https://stackoverflow.com/questions/4216985/call-to-operating-system-to-open-url
+    if sys.platform == 'win32':
+        os.startfile(url)
+    elif sys.platform == 'darwin':
+        subprocess.Popen(['open', url])
+    else:
+        try:
+            subprocess.Popen(['xdg-open', url])
+        except OSError:
+            logger.info ('Please open a browser on: ' + url)
+
+
+def inject(target, devtools=False, browser=False, timeout=None, scripts=[]):
+    timeout = time.time() + int(timeout) if timeout else 5
+
+    #
+    erb = ElectronRemoteDebugger.execute(target)
+    # launch browser?
+    if browser:
+        launch_url("http://%(host)s:%(port)s/" % erb.params)
+
+    # erb = ElectronRemoteDebugger("localhost", 8888)
+    windows_visited = set()
+    while True:
+        for w in (_ for _ in erb.windows() if _.get('id') not in windows_visited):
+            try:
+                if devtools:
+                    logger.info("injecting hotkeys script into %s" % w.get('id'))
+                    logger.debug(erb.eval(w, SCRIPT_HOTKEYS_F12_DEVTOOLS_F5_REFRESH))
+
+                for script in scripts:
+                    file = open(script)
+                    content = file.read()
+                    file.close()
+                    logger.info("injecting %s into %s" % (script, w.get('id')))
+                    logger.debug(erb.eval(w, content))
+
+            except Exception as e:
+                logger.exception(e)
+            finally:
+                # patch windows only once
+                windows_visited.add(w.get('id'))
+
+        if time.time() > timeout:
+            break
+        logger.debug("timeout not hit.")
+        time.sleep(1)
 
 
 if __name__ == "__main__":
